requirements.txt
# ================================================================================
# ⚡ FLAWLESS GENESIS ORCHESTRATOR - Dependencies
# ================================================================================
# Version: 2.0.0 LEGENDARY
# ================================================================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Async HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Redis (Async)
redis>=5.0.0

# AI Clients
anthropic>=0.18.0
openai>=1.10.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-redis>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
orjson>=3.9.0

# VORTEX v2.1 Video Assembly Dependencies
sse-starlette>=1.8.2

# RAGNAROK v4.0 Enhancement Agents
numpy>=1.24.0

# RAGNAROK v7.0 APEX Legendary Agents (7.5-15)
scipy>=1.10.0

# VORTEX v7.5 Post-Production Agents (7.75, 6.5, 7.25)
# Uses numpy (above), pydantic (above), asyncio (stdlib)
# Optional: opencv-python>=4.8.0 for advanced video analysis

# Qdrant Vector Store
qdrant-client>=1.7.0

# LangGraph Orchestration (for video brief intake + V2 Creative Director)
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-anthropic>=0.1.0

# Agent 7: Video Assembly + R2 Storage
boto3>=1.34.0
botocore>=1.34.0

# Stripe Webhook Integration
stripe>=8.0.0

# Nexus Orchestrator - Customer Lifecycle
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aio-pika>=9.3.0

# Supabase Token System
supabase>=2.0.0

# NEXUS Brain Knowledge System
pyyaml>=6.0.0
